Add Z Information, UAS Info Object
The edits propose that altitude values are only allowed / used on LineString geographies, and that the maxAltitude value of FlightPart be at least as high as the highest point in the LineString. We also tied the datum of the geography to that of the maxAltitude object; maxAltitude defines the datum for both. Technically coordinate reference systems like WGS84 do define elevation datums, but for the convenience of localized UAS operations it seems more efficient to override that and use the datum defined in the Flight Declaration.

We'd also like to start a discussion on multiple datums in the altitude object. This provides flexibility but may complicate future routing and deconfliction efforts; if several drones operating in the same area are using different datums, some translation is involved to deconflict them. Particularly, working from AGL, each operation effectively sets its own zero point based on local topography. AGL is the datum of choice for low altitude operations and we are not suggesting to change that, or to restrict the allowable altitude datums. However, we should give some thought to making a variety of datums play well together once we are deconflicting and routing on the exchange.